Add files via upload
This is a sequential deep learning resume model that will calssify whether applicants to a data analytics program are qualified or not. The model was ran using synthetic data that is at the top of the document that was built with the faker package. The model has an accuracy of 87%. This will likely change with real data. However, for demonstration purposes, we felt that this was a good representation of how a deep learning resume scanning model could work in real life.